[FIX] hr_timesheet: change department
- Set an employee in a department A
- Create some timesheet entries for this employee, validate them
- Change the employee to department B

An error arise:
> You cannot modify an entry in a confirmed timesheet.

The field `department_id` on the `account.analytic.line` is a related on
`user.employee_ids.department_id`. By changing the department on the
employee, it changes the department on the analytic line (= timesheet).

There is no real need of a related field. Actually, it makes more sense
to change the department with the user (e.g. at creation or change of
the user), and then do not change it later on. Indeed, a timesheet entry
that was created when the employee was in department A should remain in
department A even if the employee change of department.

opw-1824913