[FIX] l10n_es: Avoid account.invoice.tax fragmentation on "IVA exento"
Steps to reproduce:

- Create a vendor bill
- Add a line with tax "IVA Soportado exento (operaciones corrientes)" and an analytic
  account.
- Add another line with same tax but different analytic account.

Current behavior:

There are 2 account.invoice.tax for the tax, one per analytic account

Expected behavior:

There should be only one.

This is important because some developments like SII sendings to fiscal authority
depend on these lines.

The reason behind the problem is that the flag "Include in Analytic Cost" is ignored
when there are no configured accounts for the tax. We may want to fix this behavior
instead, but it's debatable which one is the suitable one.

As for these cases, it's clear that we can set a tax account as there's no tax at
all (they are 0% percentage), this workaround is the best option.

closes odoo/odoo#48041

X-original-commit: c2b1d90e250a495cbe57695150eaad51457183de
Signed-off-by: Josse Colpaert <jco@openerp.com>