[FIX] base: do not attempt to upgrade uninstallable addons
Assume you have a database with modules A and B installed.
Assume that A depends on B.
Assume that B is installable and A is not installable
(i.e. A has installable=False in its manifest).

When attempting to upgrade B, Odoo fails because it also attemps to upgrade A which is not installable.

This scenario happens when doing incremental version migration.
We want to be able to upgrade lower level addons and
test them with real data before migrating higher level addons.