[ADD] l10n_ar: generic changes to use in QR code
AFIP give us new resolution 4291 that request to print QR code in
printed version of electronic invoices. For that we:

1) update the xml report footer style
2) add generic method to sanitze the identification number we are
sending, the sanitize method is added here in order to be re used for
other functionalities across electronic invoice and reports. If we are
not able to sanitize the number then we show an error to the user
3) update translations to add new raise message translated term

closes odoo/odoo#65032

Related: odoo/enterprise#15947
Signed-off-by: Josse Colpaert <jco@openerp.com>