[FIX] pos_cash_rounding: fix cash rounding for cash method only
- Go to Point of Sale > Configuration > Settings & enable "Cash Rounding"
- Configure POS:
  * Enable Invoicing
  * Enable "Cash Rounding" and "Only on cash methods"
  * Select 0.05 HALF-UP Rounding Method
- Start POS session
- Select a Product and change its Price to a value that should be rounded (i.e. $ 16.51)
- Select a Customer
- Proceed to Payment
- Enable Invoice
- Select Bank Payment method
- Validate
An error is raised: "Could not fully process the POS Order: Order / is not fully paid.".
And a popup is diplayed: "Please print the invoice from backend"

The issue comes from the fact that amount is also rounded for other payment methods
than cash, even if "Only on cash methods" has been enabled.

opw-2449222

closes odoo/odoo#65297

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>