[FIX] stock: Allows to transfer partially with cancel
Do a transfer in two steps (Vendors > Input > Stock)
with propagate_cancel moves == True.

On the first picking, do a partial transfer in terms of quantities
(e.g.: 8 instead of 10) and do not request a backorder.

The remaining quantities are cancelled either on the first picking
either on the second one.