[FIX] event_sale: remove 'need to be paid' banner for free ticket
This commit fixes an issue where if a ticket is free, the registration would
still "need to be paid" by the attendee.

This behavior is contradictory for a free ticket, so the attendee now has to
pay the ticket if it is not paid and is not a free ticket.

task-3012928

closes odoo/odoo#102996

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>