Init comodel_name on relational fields from generic type's arg
This change allows to take advantage of field descriptors on relational fields to initialize the comodel_name from optional type's arg that can be specified on field declaration. With this change, this kind of syntax becomes valid and will avoid errors in the comodel_name args.

partner = fields.Many2one[Partner]()

This code is not part of the _Relational class constructor since we need to get access to the __orig_class__ attribute which is set on the instance by the Generic metaclass after the call to the _Relational class constructor.