[FIX] sms: invisible description when empty
Before this change, when the recipient_single_description field was empty (the field is then false), the field box was still displayed before the number. Since the field is not editable there is just a blank space.

This PR hides the field when it is empty to avoid having the empty space.

closes odoo/odoo#105436

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>