[FIX] website_crm_partner_assign: prevent opening partnership analysis tree/form views
on clicking the graph view of the Partnership Analysis, currently the tree
view and form view is opened. actually the tree and form view for this model
are not defined in the code and thus the end user get the tree view with only
ID field in it.

As those view have no sense, prevent from jumping on those.

closes odoo/odoo#106375

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>