[FIX] dx: port back js tooling gitignore properties
The tooling has been introduce in later version of odoo.
It is a problem when somebody switches to older branches without
disabling the tooling as it would then try to commit the tooling config
files.

We just add a few lines in the gitignore to prevent this issue.

closes odoo/odoo#106878

Related: odoo/enterprise#34518
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>