[FIX] payment_stripe: Enable button after card is marked as invalid to continue workflow
When the stripe card is invalid we break the `PayEvent` workflow to prevent purchases,
the problem is that the button is never enabled again.

This gives the wrong impression that something is happening because the payment button
is disabled and the spinning icon is there, by `enabling` the button again we ensure
that if the card is invalid the user changes its values and tries again.

closes odoo/odoo#107446

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>