[FIX] snailmail: Fix base snailmail layouts
Due to some changes to the layout and to some upcoming changes to the
service provider's validation area, the layout had to be fixed to avoid
validation errors. This commit also patches the styling due to a bug
messing up the invoice when "Customer Addresses" is enabled in the
settings.

closes odoo/odoo#107870

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>