[FIX] point_of_sale: prevent JSON modification in `init_from_JSON`
Before this commit: after calling the `init_from_JSON` of the order line,
the `pack_lot_ids` would be changed. It will cause a problem if this
function is called before sending the JSON to the backend or saving it.

The solution is to copy its value and prevent it from changing.

opw-3088922

closes odoo/odoo#108214

Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>