[FIX] website_mail_channel: stop discussion block disappearing on move
Before this PR, when moving the snippet's position, `d-none` gets added
which should not.
With this PR, the `cleanForSave` method is removed as it applied that
`d-none` and the logic is added to the `destroy` method of the public
widget (on page load, we make the snippet visible when discussion group
info is processed and so it is logical to hide it again in `destroy`).

task-3107451

closes odoo/odoo#108790

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>