[FIX] point_of_sale: allow to close order after validation before printing
Before this commit: if "Automatic Receipt Printing" is enabled in a PoS
restaurant, and a printer is configured, an error will raise if the user clicks
on "New Order" before the end of the printing process.

To reproduce the issue, I added a sleep at the beginning of the
`_printReceipt` of `AbstractReceiptScreen`, and clicked on "New Order"
immediately after validating the order.

The solution is to keep the `currentOrder` in a variable before printing.

opw-3096677

closes odoo/odoo#111041

Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>