[FIX] calendar-sms : remove broken button
This button has been broken since the start of V14.
As there is no clean way to fix this, it was decided to remove it
instead.
This change will not affect existing V14 DB unless they specificaly
update the XML or the module.

This button was broken because the sms.composer was called
with the res.model/res.id of the unique calendar event that
was active. But the overwrite of `_sms_get_default_partners`
would then return the res.ids of all the unconfirmed atendees'
partner id.

This would cause an issue a this fetch to default partner could
return multiple ids when the base id only add one and thus had called
`_compute_recipient_single` instead `_compute_recipients`

task-id : 3162304

closes odoo/odoo#111662

Signed-off-by: Arnaud Joset <arj@odoo.com>