[FIX] hw_drivers: Avoid duplicated actions execution
Before this commit:
Assuming we have a "Shipping Labels Printer" set on an operation type.
Assuming we also have a delivery carrier which send Label through the chatter of this picking type (like DHL or BPost).
If we have several sessions connected with the same user (author of the chatter message), for examples, on different devices.
Each session will send the same IoT request to print the label (as the code rely on the bus).

In consequences, the same label will be printed multiple times

After this commit:
The label is print only once

Associated enterprise PR: https://github.com/odoo/enterprise/pull/36904

opw-3081423

closes odoo/odoo#112355

Related: odoo/enterprise#36904
Signed-off-by: Sens Loan (lse) <lse@odoo.com>