[FIX] web_editor: fix shapes that disappear
This commit fixes a bug where in some cases a shape disappears after
applying a flip to it. This bug has been present since commit [1].

Steps to reproduce the bug:

- Choose the "Clean Theme" for a website.
- Drag and drop a "Call to action" snippet on the page.
- Click on one of the two buttons to flip the shape of this snippet.
- Bug => The shape disappears.

This was due to shape colors set by theme variables in some theme
snippets (e.g. 'o-color-1'). When the background image url was recreated
to apply the "flip", the color was not converted to hexadecimal.

[1]: 42b3ad1

task-2824607
opw-3069213
opw-3057533

closes odoo/odoo#114684

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>