[FIX] account: fix ir.rule on tax repartition lines
Before this commit, you could search `account.tax.repartition.line`
and get records from all the existing companies,
which is not supposed to happen.

It should only return records of the `active_company_ids`.

That was because there were no `ir.rule` preventing that.
This fixes the issue.

closes odoo/odoo#114718

Related: odoo/enterprise#37931
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>