[FIX] account: no default taxes on accounts
So fat, the default sale and purchase taxes were so on the
account.accounts depending on the user type id.
These taxes were meant to be used as default taxes on products
and not accounts, so this is not right.
It also is an issue since the default tax isn't visible
on the COA list until the user add it from the optional fields.

Task id # 3223731

closes odoo/odoo#114892

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>