[FIX] mail_client_extension/mail_plugin: let users allow mail plugin to work
This commit fixes a regression introduced in https://github.com/odoo/odoo/pull/104323.

That fix was preventing the default behavior of the 1st `.oe_login_form button[type=submit]` found.

It turns out that, for allowing the mail plugin to work, you actually need to have the default behavior of that element working normally.

Without this patch, the `&do=1` query is never added to the POST parameters. Thus, the user is unable to authorize their mail plugin.

@moduon MT-2552 MT-2553 OPW-3230686

closes odoo/odoo#115198

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>