[FIX] website: neutralize sub-pixel height difference in images wall
Inside the Images Wall snippet, images are dispatched to columns
depending on the height already reached by each column. This computation
relies on a sub-pixel height which leads to a confusing behavior.

This commit rounds the sub-pixel height to a visible pixel height to
avoid the confusion.

Steps to reproduce:
In the default images of the Images Wall snippet, the third image (sign)
is one pixel taller than the other ones.
- Drop an Images Wall snippet.
- In 14.0, upload the default images from v15.0.
- Select the "sign" image.
- Move it to the previous position.

=> The sixth image (wine glass) moved as well.

task-2990053

Part-of: odoo/odoo#120523