[FIX] survey: use bg-success only if the answer is correct
Before this commit:
In MCQ-type questions, the answers on the 'Review my answers' page were
displayed with a green background if their score was above 0, regardless of
whether they were correct. Otherwise, answers were displayed with a red
background.

After this commit:
The answers will be shown based on whether they are correct or not, regardless
of their score. I.e. correct answers will be displayed with a green background,
and all other answers will be displayed with a red background.

Task-3252605

closes odoo/odoo#123878

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>