[FIX] website_crm, website_form: prevent remove model required fields
Depending on the modules installed, the action of the form on the
`/contactus` page changes. For example, if you have `website_crm`
installed, the action will be set to "Create an Opportunity" but if you
only have `website_form` installed, the action will be "Send an Email".
For each action, there are mandatory fields that cannot be deleted by
the user. Unfortunately, the templates for the `/contactus` forms didn't
have the right fields marked as mandatory. This led to the following bug
(with only `website_form` installed):
- Go to `/contactus`
- Edit the page

=> The Email and Subject fields can be deleted. However, when you drop a
form and set the action to "Send an Email", these fields are mandatory
and cannot be deleted. This commit fixes this bug by ensuring that these
fields are marked as mandatory on `/contactus`. This commit fixes the
same issue with the `website_crm` module.

task-3302433

closes odoo/odoo#122858

Signed-off-by: Romain Derie (rde) <rde@odoo.com>