[IMP] stock: Improves product.action_open_quants performances
Before this change, when the action was called the query to merge duplicate quant entries was called on all the products. This could lead to severe performance issue on large database. With this change, the query searching for duplicate quants only applies on quants for the product ids the action is launched for